Split analysis app directory overrides
